requirements.txt
# E-ink Web UI Requirements
# Core web framework
Flask>=2.3.0

# Image processing
Pillow>=9.0.0

# Scientific computing for image arrays
numpy>=1.21.0

# QR code generation
qrcode[pil]>=7.0.0

# Optional: For better development experience
# gunicorn>=20.1.0  # Production WSGI server
# python-dotenv>=0.19.0  # Environment variable management